AP_Math: AP_GeodesicGrid: use uint8_t for _neighbor_umbrellas
All integers there are limited to the range [0,20), so uint8_t is enough.

+++ libraries/AP_Math/AP_GeodesicGrid.h
@@ -202,18 +202,18 @@ private:
          * The umbrella's components. The value of #components[i] is the
          * icosahedron triangle index of the i-th component.
          */
-        int components[5];
+        uint8_t components[5];
         /**
          * The fields with name in the format vi_cj are interpreted as the
          * following: vi_cj is the index of the vector, in the icosahedron
          * triangle pointed by #components[j], that matches the umbrella's i-th
          * vertex.
          */
-        int v0_c0;
-        int v1_c1;
-        int v2_c1;
-        int v4_c4;
-        int v0_c4;
+        uint8_t v0_c0;
+        uint8_t v1_c1;
+        uint8_t v2_c1;
+        uint8_t v4_c4;
+        uint8_t v0_c4;
     } _neighbor_umbrellas[6];
 
     /**
